+++ src/app/(team)/create/page.tsx
@@ -1,56 +1,21 @@
-"use client"
+import prisma from "@/db/prisma"
+import CreateForm from "@/components/teams/create-form"
+import { getServerSession } from "next-auth"
+import option from "@/app/api/auth/[...nextauth]/options"
 
-import { useForm, SubmitHandler } from "react-hook-form"
-import { InputFields } from "@/libs/schema/teamSchema"
-import { useRouter } from "next/navigation"
+import { redirect } from "next/navigation"
 
-export default function CreateTeamPage() {
+export default async function CreateTeamPage() {
   const {
-    register,
-    handleSubmit,
-    setError,
-    formState: { errors }
-  } = useForm<InputFields>()
-  const router = useRouter()
+    user: { email }
+  } = await getServerSession(option)
+  const { id } = await prisma.team.findFirst({
+    where: { members: { some: { user: { email } } } }
+  })
 
-  const submit: SubmitHandler<InputFields> = async data => {
-    const request = await fetch("/api/team/create", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify(data)
-    })
-
-    if (!request.ok) {
-      const response = await request.json()
-      response.errors.map(e =>
-        setError(e.name, { type: "manual", message: e.message })
-      )
-    }
-
-    const { id } = await request.json()
-    router.push(`/${id}`)
+  if (id) {
+    redirect(`/${id}`)
   }
 
-  return (
-    <div>
-      <h1>New Team</h1>
-      <form onSubmit={handleSubmit(submit)}>
-        <div>
-          <label htmlFor="team-name">Name</label>
-          <input
-            type="text"
-            name="team-name"
-            id="team-name"
-            {...register("name")}
-          />
-          {errors.name && (
-            <span className="text-red-500 text-sm">{errors.name.message}</span>
-          )}
-        </div>
-        <button type="submit">Create</button>
-      </form>
-    </div>
-  )
+  return <CreateForm />
 }
